+++ src/gluonts/time_feature/lag.py
@@ -33,27 +33,6 @@ from gluonts.time_feature import (
 )
 
 
-# A dictionary mapping granularity to the period length of the longest season
-# one can expect given the granularity of the time series.
-# This is similar to the frequency value in the R forecast package:
-# https://stats.stackexchange.com/questions/120806/frequency-value-for-seconds-minutes-intervals-data-in-r
-# This is useful for setting default values for past/context length for models
-# that do not do data augmentation and uses a single training example per time series in the dataset.
-FREQ_LONGEST_PERIOD_DICT = {
-    "M": 12,  # yearly seasonality
-    "W-SUN": 52,  # yearly seasonality
-    "D": 365,  # yearly seasonality
-    "B": 365,  # yearly seasonality
-    "H": 168,  # weekly seasonality
-    "T": 1440,  # daily seasonality
-}
-
-
-def longest_period_from_frequency_str(freq_str: str) -> int:
-    offset = to_offset(freq_str)
-    return FREQ_LONGEST_PERIOD_DICT[offset.name] // offset.n
-
-
 def time_features_from_frequency_str(freq_str: str) -> List[TimeFeature]:
     """
     Returns a list of time features that will be appropriate for the given frequency string.
